backend/requirements.txt
Django==3.1.2
psycopg2-binary==2.8.6
djangorestframework==3.12.1
django-cors-headers==3.5.0
django-allauth==0.43.0
dj-rest-auth==1.1.2
djangorestframework-simplejwt==4.4.0
drf-yasg==1.20.0
django-filter===2.4.0